splitting custom.d.ts in multiple files to tide up + Params=>LoadDataParameters
