+++ src/components/Entry/Summary/Wikipedia/index.tsx
@@ -6,7 +6,6 @@ import { XMLParser } from 'fast-xml-parser';
 import { createSelector } from 'reselect';
 import { format } from 'url';
 import loadData from 'higherOrder/loadData/ts';
-import { Params } from 'src/higherOrder/loadData/extract-params';
 
 import Link from 'components/generic/Link';
 import Loading from 'components/SimpleCommonComponents/Loading';
@@ -179,4 +178,4 @@ const getWikiUrl = createSelector(
     });
   },
 );
-export default loadData(getWikiUrl as Params)(Wikipedia);
+export default loadData(getWikiUrl as LoadDataParameters)(Wikipedia);
