+++ src/components/Protein/Isoforms/Selector/index.tsx
@@ -4,7 +4,6 @@ import { format } from 'url';
 import { createSelector } from 'reselect';
 
 import loadData from 'higherOrder/loadData/ts';
-import { Params } from 'higherOrder/loadData/extract-params';
 
 import descriptionToPath from 'utils/processDescription/descriptionToPath';
 import { goToCustomLocation } from 'actions/creators';
@@ -70,7 +69,7 @@ const Selector = ({
 const mapStateToProps = createSelector(
   (state: GlobalState) => state.customLocation,
   (state: GlobalState) => state.customLocation.search,
-  (customLocation, { isoform }) => ({ customLocation, isoform })
+  (customLocation, { isoform }) => ({ customLocation, isoform }),
 );
 
 const getIsoformURL = createSelector(
@@ -91,10 +90,10 @@ const getIsoformURL = createSelector(
         isoforms: '',
       },
     });
-  }
+  },
 );
 export default loadData({
   getUrl: getIsoformURL,
   mapStateToProps,
   mapDispatchToProps: { goToCustomLocation },
-} as Params)(Selector);
+} as LoadDataParameters)(Selector);
