+++ src/components/Structure/ViewerAndEntries/ProteinViewerForStructures/index.tsx
@@ -3,7 +3,6 @@ import { createSelector } from 'reselect';
 import { format } from 'url';
 
 import loadData from 'higherOrder/loadData/ts';
-import { Params } from 'higherOrder/loadData/extract-params';
 import descriptionToPath from 'utils/processDescription/descriptionToPath';
 
 import Loading from 'components/SimpleCommonComponents/Loading';
@@ -94,7 +93,7 @@ const getSecondaryStructureURL = createSelector(
 export default loadData<StructureWithSecondary, 'Secondary'>({
   propNamespace: 'Secondary',
   getUrl: getSecondaryStructureURL,
-} as Params)(
+} as LoadDataParameters)(
   loadData<PayloadList<EndpointWithMatchesPayload<EntryMetadata>>>(
     getURLForMatches,
   )(ProteinViewerForStructure),
