+++ src/types/storage.d.ts
@@ -1,0 +1,11 @@
+type IprscanMetaIDB = {
+  group: string;
+  hasResults: boolean;
+  saved: boolean;
+  localID: string;
+  remoteID: string;
+  localTitle: string;
+  status: string;
+  type: string;
+  times: Record<string, number>;
+};
